added support for srs format conversion with options (e.g. WKT ESRI etc)
